see description
Check on all operators' behavior. Fixed Conv code for group>1 that made MobileNet crash. Implemented a first version of serialization from an image to a .pb file, for both resnet and mnist (WIP). Changes in menu to manage serialization (WIP). First version of custom error management (WIP). Added a warning in the README.md for build problems depending on the chosen toolchain.