see description
Fix in MaxPool operator where some attributes were not correctly initialized. Fix in console clearing. Fix in dataset folder name input that is now trimmed. Fix in DirEntries for serialization because files (not dirs) in numerical format were taken too, so .is_dir() is used to check if the entry is a folder. Major code refactor.