see description
Added images for README.md and completed documentation on README.md. Added protoc_generated folder to .gitignore. Standardized custom datasets to include only images from Imagenette repo. Fixed conflicting binary and library names. Fixed maxpool operator and made it more similar to the Python ONNX implementation. Added OnnxErrors in operators' implementation. Added error to the Reshape operator for allow_zero equal to 1 and both a zero and a -1 are in the target shape, which means that the target shape cannot be determined. Minor fixes and code refactor.